[TNZ-27070]: Fix UAA on standard ports
