+++ server/src/test/java/org/cloudfoundry/identity/uaa/provider/saml/SamlRedirectUtilsTest.java
@@ -73,112 +73,4 @@ class SamlRedirectUtilsTest {
         assertThat(SamlRedirectUtils.getZonifiedEntityId("local-name", newZone)).isEqualTo("new-zone-id.local-name");
     }
 
-    @Test
-    void normalizeUrlForPortComparison_withHttpStandardPort_removesPort() {
-        String urlWithPort = "http://example.com:80/path";
-        String expected = "http://example.com/path";
-        
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(urlWithPort);
-        
-        assertThat(result).isEqualTo(expected);
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withHttpsStandardPort_removesPort() {
-        String urlWithPort = "https://example.com:443/path";
-        String expected = "https://example.com/path";
-        
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(urlWithPort);
-        
-        assertThat(result).isEqualTo(expected);
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withHttpNonStandardPort_keepsPort() {
-        String urlWithPort = "http://example.com:8080/path";
-        
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(urlWithPort);
-        
-        assertThat(result).isEqualTo(urlWithPort);
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withHttpsNonStandardPort_keepsPort() {
-        String urlWithPort = "https://example.com:8443/path";
-        
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(urlWithPort);
-        
-        assertThat(result).isEqualTo(urlWithPort);
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withoutPort_returnsUnchanged() {
-        String urlWithoutPort = "http://example.com/path";
-        
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(urlWithoutPort);
-        
-        assertThat(result).isEqualTo(urlWithoutPort);
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withQueryAndFragment_preservesThem() {
-        String urlWithPort = "http://example.com:80/path?query=value#fragment";
-        String expected = "http://example.com/path?query=value#fragment";
-        
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(urlWithPort);
-        
-        assertThat(result).isEqualTo(expected);
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withUserInfo_preservesIt() {
-        String urlWithPort = "http://user:pass@example.com:80/path";
-        String expected = "http://user:pass@example.com/path";
-        
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(urlWithPort);
-        
-        assertThat(result).isEqualTo(expected);
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withNullUrl_returnsNull() {
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(null);
-        
-        assertThat(result).isNull();
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withMalformedUrl_returnsOriginal() {
-        String malformedUrl = "not-a-valid-url";
-        
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(malformedUrl);
-        
-        assertThat(result).isEqualTo(malformedUrl);
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withDifferentScheme_keepsPort() {
-        String urlWithPort = "ftp://example.com:80/path";
-        
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(urlWithPort);
-        
-        assertThat(result).isEqualTo(urlWithPort);
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withEmptyString_returnsEmpty() {
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison("");
-        
-        assertThat(result).isEqualTo("");
-    }
-
-    @Test
-    void normalizeUrlForPortComparison_withComplexUrl_handlesCorrectly() {
-        String urlWithPort = "https://subdomain.example.com:443/saml/SSO/alias/provider-name?RelayState=https://app.example.com&param=value#section";
-        String expected = "https://subdomain.example.com/saml/SSO/alias/provider-name?RelayState=https://app.example.com&param=value#section";
-        
-        String result = SamlRedirectUtils.normalizeUrlForPortComparison(urlWithPort);
-        
-        assertThat(result).isEqualTo(expected);
-    }
 }
